+++ history.md
@@ -1,5 +1,6 @@
 ### 2.1.20-beta
 - complex-func:getRandomData函数同步调整
+- IdData:id类型函数BUG修复，restart模式长度判断缺失，maxAction函数变量变更不生效，优化实现
 
 ### 2.1.19
 - 依赖版本修正